Found 3d tables for timing
Found the threeD tables used for timing, and the algorithm used to read
the tables and perform the bilinear interpolation. as usual there is one
or two tidbits that while they are clear, they simply dont make sense.

also updated pcm pinout to reflect that the pictures are the picture of
the harness connectors, not the PCM. I'd known this for a while :)